+++ metagpt/reflect/__init__.py
@@ -1,3 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # @Desc   : reflection module
+
+from metagpt.reflect import agent_reflect
+
+__all__ = [
+    "agent_reflect",
+    "LongTermMemory",
+]
