--- metagpt/reflect_and_retrieve/Prompt_template/poignancy_chat_v1.txt
@@ -1,17 +1,0 @@
-poignancy_chat_v1.txt
-
-!<INPUT 1>!: agent name
-!<INPUT 1>!: iss
-!<INPUT 2>!: name 
-!<INPUT 3>!: event description
-
-<commentblockmarker>###</commentblockmarker>
-Here is a brief description of !<INPUT 0>!. 
-!<INPUT 1>!
-
-On the scale of 1 to 10, where 1 is purely mundane (e.g., routine morning greetings) and 10 is extremely poignant (e.g., a conversation about breaking up, a fight), rate the likely poignancy of the following conversation for !<INPUT 2>!.
-
-Conversation: 
-!<INPUT 3>!
-
-Rate (return a number between 1 to 10):
