--- metagpt/reflect_and_retrieve/pycodetester.py
@@ -1,3 +1,0 @@
-import pycodestyle as pcs
-checker = pcs.StyleGuide()
-checker.input_dir('./')
